refactor out cli::parse_args
